feat: APIManager protocol로 변경
- APIManager<class -> protocol>
- 코드 반복되는 것을 수정

#2
